tests: OASIS C_Verify() focus on RSA_PKCS
Let's avoid any confusion for this test: let's sign and then verify
without any SHA256.

Moreover the C_Logout() is performed right after the signature in order
to highlight that it is not required to be login-ed when a C_Verify()
is performed.

The CKA_ID is not harded coded because it may change between various
PKCS11 systems. I'd prefer to record it as commented code in ordre
to provide some samples for this argument that requires an hex2bin().